fix inheriting projects
